+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/ProfileTest.java
@@ -1,24 +1,47 @@
 package guru.qa.niffler.test.web;
 
-import guru.qa.niffler.jupiter.extension.UsersQueueExtension;
+import guru.qa.niffler.config.Config;
+import guru.qa.niffler.jupiter.annotation.Category;
+import guru.qa.niffler.model.CategoryJson;
+import guru.qa.niffler.page.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 
-import static guru.qa.niffler.jupiter.extension.UsersQueueExtension.*;
+import static com.codeborne.selenide.Selenide.*;
 
-@ExtendWith(UsersQueueExtension.class)
 public class ProfileTest {
 
+  private static final Config CFG = Config.getInstance();
+
+  @Category(
+      username = "Oleg"
+  )
   @Test
-  void testWithEmptyUser0(@UserType(empty = true) StaticUser user) throws InterruptedException {
-    Thread.sleep(1000);
-    System.out.println(user);
+  void archivedCategoryShouldNotPresentInCategoriesList(CategoryJson category) {
+    open(CFG.frontUrl(), LoginPage.class)
+        .login("Oleg", "12345")
+        .getTopMenu()
+        .goToProfile()
+        .clickArchiveCategoryByName(category.name())
+        .clickArchiveButton()
+        .shouldBeVisibleSuccessMessage(category.name(), "is archived")
+        .shouldNotBeVisibleCategory(category.name());
   }
 
+  @Category(
+      username = "Oleg",
+      archived = true
+  )
   @Test
-  void testWithEmptyUser1(@UserType(empty = true) StaticUser user) throws InterruptedException {
-    Thread.sleep(1000);
-    System.out.println(user);
+  void activeCategoryShouldPresentInCategoriesList(CategoryJson category) {
+    open(CFG.frontUrl(), LoginPage.class)
+        .login("Oleg", "12345")
+        .getTopMenu()
+        .goToProfile()
+        .clickShowArchiveCategoryButton()
+        .clickUnarchiveCategoryByName(category.name())
+        .clickUnarchiveButton()
+        .shouldBeVisibleSuccessMessage(category.name(), "is unarchived")
+        .clickShowArchiveCategoryButton()
+        .shouldBeVisibleCategory(category.name());
   }
-
 }
